+++ src/main/java/de/kamillionlabs/hateoflux/assembler/SealedResourceListAssemblerModule.java
@@ -19,8 +19,10 @@
 package de.kamillionlabs.hateoflux.assembler;
 
 import de.kamillionlabs.hateoflux.model.hal.HalListWrapper;
+import de.kamillionlabs.hateoflux.model.hal.HalPageInfo;
 import de.kamillionlabs.hateoflux.model.hal.Relation;
 import de.kamillionlabs.hateoflux.model.link.Link;
+import de.kamillionlabs.hateoflux.utility.SortCriteria;
 import org.springframework.lang.NonNull;
 import org.springframework.web.server.ServerWebExchange;
 
@@ -59,6 +61,35 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
     }
 
     /**
+     * Creates a paged empty {@link HalListWrapper} including hypermedia links applicable to the entire list.
+     *
+     * @param listItemTypeAsNameOrigin
+     *         the class of the resource from which the list name is derived, typically pluralized to represent both
+     *         resource and embedded resource types (also see {@link Relation})
+     * @param pageSize
+     *         the requested/max number of elements in a single page
+     * @param pageNumber
+     *         the current page number
+     * @param sortCriteria
+     *         sort criteria (property and direction) of the page
+     * @param exchange
+     *         provides the context of the current web exchange, such as the base URL
+     * @return an initialized {@link HalListWrapper} with relevant hypermedia links for the entire list
+     *
+     * @see #createEmptyListWrapper(String, ServerWebExchange)
+     */
+    default HalListWrapper<ResourceT, EmbeddedT> createEmptyListWrapper(@NonNull Class<?> listItemTypeAsNameOrigin,
+                                                                        int pageSize,
+                                                                        int pageNumber,
+                                                                        List<SortCriteria> sortCriteria,
+                                                                        ServerWebExchange exchange) {
+        HalListWrapper<ResourceT, EmbeddedT> emptyWrapper = HalListWrapper.empty(listItemTypeAsNameOrigin);
+        HalPageInfo halPageInfo = HalPageInfo.assembleWithPageNumber(pageSize, 0, pageNumber);
+        return emptyWrapper.withLinks(buildLinksForResourceList(halPageInfo, sortCriteria, exchange))
+                .withPageInfo(halPageInfo);
+    }
+
+    /**
      * Creates an empty {@link HalListWrapper} including hypermedia links applicable to the entire list.
      *
      * @param listName
@@ -76,10 +107,36 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
     }
 
     /**
+     * Creates an empty {@link HalListWrapper} including hypermedia links applicable to the entire list.
+     *
+     * @param listName
+     *         the given name for the list
+     * @param pageSize
+     *         the requested/max number of elements in a single page
+     * @param pageNumber
+     *         the current page number
+     * @param sortCriteria
+     *         sort criteria (property and direction) of the page
+     * @param exchange
+     *         provides the context of the current web exchange, such as the base URL
+     * @return an initialized {@link HalListWrapper} with relevant hypermedia links for the entire list
+     *
+     * @see #createEmptyListWrapper(Class, ServerWebExchange)
+     */
+    default HalListWrapper<ResourceT, EmbeddedT> createEmptyListWrapper(@NonNull String listName,
+                                                                        int pageSize,
+                                                                        int pageNumber,
+                                                                        List<SortCriteria> sortCriteria,
+                                                                        ServerWebExchange exchange) {
+        HalListWrapper<ResourceT, EmbeddedT> emptyWrapper = HalListWrapper.empty(listName);
+        HalPageInfo halPageInfo = HalPageInfo.assembleWithPageNumber(pageSize, 0, pageNumber);
+        return emptyWrapper.withLinks(buildLinksForResourceList(halPageInfo, sortCriteria, exchange));
+    }
+
+    /**
      * Main method for building all links for a list of resources and embedded resources, including a self-link and
-     * other
-     * contextual links applicable to the entire list. It aggregates results from {@link #buildSelfLinkForResourceList}
-     * and {@link #buildOtherLinksForResourceList}.
+     * other  contextual links applicable to the entire list. It aggregates results from
+     * {@link #buildSelfLinkForResourceList} and {@link #buildOtherLinksForResourceList}.
      *
      * @param exchange
      *         provides the context of the current web exchange, such as the base URL
@@ -89,6 +146,36 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
     default List<Link> buildLinksForResourceList(ServerWebExchange exchange) {
         List<Link> links = new ArrayList<>();
         links.add(buildSelfLinkForResourceList(exchange).withSelfRel());
+        links.addAll(buildOtherLinksForResourceList(exchange));
+        return links;
+    }
+
+    /**
+     * Main method for building all links for a paged list of resources and embedded resources, including
+     * navigational links and other contextual links applicable to the entire list. It aggregates results from
+     * {@link #buildSelfLinkForResourceList} and {@link #buildOtherLinksForResourceList}.
+     *
+     * @param pageInfo
+     *         pagination information about the request page
+     * @param sortCriteria
+     *         sort criteria (property and direction) of the page
+     * @param exchange
+     *         provides the context of the current web exchange, such as the base URL
+     * @return a list of {@link Link} objects representing hypermedia links for the entire list of resource and embedded
+     * resource types
+     */
+    default List<Link> buildLinksForResourceList(HalPageInfo pageInfo,
+                                                 List<SortCriteria> sortCriteria,
+                                                 ServerWebExchange exchange) {
+        List<Link> links = new ArrayList<>();
+        Link baseLink = buildSelfLinkForResourceList(exchange).withSelfRel();
+        List<Link> navigationLinks;
+        if (sortCriteria == null) {
+            navigationLinks = baseLink.deriveNavigationLinks(pageInfo);
+        } else {
+            navigationLinks = baseLink.deriveNavigationLinks(pageInfo, sortCriteria);
+        }
+        links.addAll(navigationLinks);
         links.addAll(buildOtherLinksForResourceList(exchange));
         return links;
     }